# src/routes/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "Record_string.any_": {
        "properties": {},
        "additionalProperties": {},
        "type": "object",
        "description": "Construct a type with a set of properties K of type T"
      },
...
      "Record_string.unknown_": {
        "properties": {},
        "additionalProperties": {},
        "type": "object",
        "description": "Construct a type with a set of properties K of type T"
      },
      "DidOperation.Create": {
        "enum": [
          "createDID"
        ],
        "type": "string"
      },
      "CreateDidOperationOptions": {
        "properties": {
          "serviceEndpoint": {
            "type": "string"
          },
          "operation": {
            "$ref": "#/components/schemas/DidOperation.Create"
          }
        },
        "required": [
          "operation"
        ],
        "type": "object"
      },
      "DidOperation.Update": {
        "enum": [
          "updateDIDDoc"
        ],
        "type": "string"
      },
      "DidDocument": {
        "$ref": "#/components/schemas/Record_string.any_"
      },
      "UpdateDidOperationOptions": {
        "properties": {
          "did": {
            "type": "string"
          },
          "didDocument": {
            "$ref": "#/components/schemas/DidDocument"
          },
          "operation": {
            "$ref": "#/components/schemas/DidOperation.Update"
          }
        },
        "required": [
          "did",
          "didDocument",
          "operation"
        ],
        "type": "object"
      },
      "DidOperation.Deactivate": {
        "enum": [
          "deactivate"
        ],
        "type": "string"
      },
      "DeactivateDidOperationOptions": {
        "properties": {
          "did": {
            "type": "string"
          },
          "operation": {
            "$ref": "#/components/schemas/DidOperation.Deactivate"
          }
        },
        "required": [
          "did",
          "operation"
        ],
        "type": "object"
      },
      "DidOperation.AddResource": {
        "enum": [
          "addResource"
        ],
        "type": "string"
      },
      "AddResourceDidOperationOptions": {
        "properties": {
          "did": {
            "type": "string"
          },
          "resource": {
            "additionalProperties": false,
            "type": "object"
          },
          "resourceId": {
            "type": "string"
          },
          "operation": {
            "$ref": "#/components/schemas/DidOperation.AddResource"
          }
        },
        "required": [
          "did",
          "resource",
          "resourceId",
          "operation"
        ],
        "type": "object"
      },
      "DidOperationOptions": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/CreateDidOperationOptions"
          },
          {
            "$ref": "#/components/schemas/UpdateDidOperationOptions"
          },
          {
            "$ref": "#/components/schemas/DeactivateDidOperationOptions"
          },
          {
            "$ref": "#/components/schemas/AddResourceDidOperationOptions"
          }
        ]
      },
...
      "SchemaId": {
        "type": "string",
        "example": "WgWxqztrNooG92RXvxSTWv:2:schema_name:1.0"
      },
      "CreateSchemaInput": {
        "properties": {
          "issuerId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "attributes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "endorse": {
            "type": "boolean"
          },
          "endorserDid": {
            "type": "string"
          }
        },
        "required": [
          "issuerId",
          "name",
          "version",
          "attributes"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      },
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "credo-controller",
    "version": "2.0.0",
    "description": "Rest endpoint wrapper for using your agent over HTTP",
    "license": {
      "name": "Apache-2.0"
    },
    "contact": {}
  },
  "paths": {
...
    "/polygon/create-keys": {
      "post": {
        "operationId": "CreateKeyPair",
        "responses": {
          "200": {
            "description": "Secp256k1KeyPair",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "address": {
                      "type": "string"
                    },
                    "publicKeyBase58": {
                      "type": "string"
                    },
                    "privateKey": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "address",
                    "publicKeyBase58",
                    "privateKey"
                  ],
                  "type": "object"
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Create Secp256k1 key pair for polygon DID",
        "tags": [
          "Polygon"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated",
              "Basewallet"
            ]
          }
        ],
        "parameters": []
      }
    },
    "/polygon/create-schema": {
      "post": {
        "operationId": "CreateSchema",
        "responses": {
          "200": {
            "description": "Schema JSON",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Create polygon based W3C schema",
        "tags": [
          "Polygon"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "schema": {
                    "$ref": "#/components/schemas/Record_string.unknown_"
                  },
                  "schemaName": {
                    "type": "string"
                  },
                  "did": {
                    "type": "string"
                  }
                },
                "required": [
                  "schema",
                  "schemaName",
                  "did"
                ],
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/polygon/estimate-transaction": {
      "post": {
        "operationId": "EstimateTransaction",
        "responses": {
          "200": {
            "description": "Transaction Object",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Estimate transaction",
        "tags": [
          "Polygon"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated",
              "Basewallet"
            ]
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DidOperationOptions"
              }
            }
          }
        }
      }
    },
    "/polygon/{did}/{schemaId}": {
      "get": {
        "operationId": "GetSchemaById",
        "responses": {
          "200": {
            "description": "Schema Object",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Fetch schema details",
        "tags": [
          "Polygon"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "did",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/anoncreds/schemas/{schemaId}": {
      "get": {
        "operationId": "GetSchemaById",
        "responses": {
          "200": {
            "description": "get schema by Id",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "type": "object"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "ver": "1.0",
                      "id": "WgWxqztrNooG92RXvxSTWv:2:schema_name:1.0",
                      "name": "schema",
                      "version": "1.0",
                      "attrNames": [
                        "string"
                      ],
                      "seqNo": 351936
                    }
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Get schema by schemaId",
        "tags": [
          "AnonCreds - Schemas (Preserved)"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SchemaId"
            }
          }
        ]
      }
    },
    "/anoncreds/schemas": {
      "post": {
        "operationId": "CreateSchema",
        "responses": {
          "200": {
            "description": "get schema",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "type": "object"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "state": "finished",
                      "schema": {
                        "issuerId": "did:key:z6MkpTHR8VNsBxYAAWHut2Geadd9jSwuBV8xRoAnwWsdvktH",
                        "name": "Test Schema",
                        "version": "1.0.0",
                        "attrNames": [
                          "Name",
                          "Age"
                        ]
                      },
                      "schemaId": "LRCUFcizUL74AGgLqdJHK7:2:Test Schema:1.0.0"
                    }
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Create schema",
        "tags": [
          "AnonCreds - Schemas (Preserved)"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchemaInput"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}